requirements.txt
Flask==2.2.5
Flask-SQLAlchemy>=3.1,<4
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.0
stripe>=10.0.0
Flask-Migrate==4.0.5
alembic>=1.13.0
Weasyprint
PyPDF2>=3.0,<4
itsdangerous
Flask-Mail>=0.9.1
requests>=2.31
APScheduler==3.10.4
openai>=1.45.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
opencv-python-headless>=4.10.0.84
pydantic>=2.6.4
tenacity>=8.2.3
pytesseract>=0.3.10
numpy>=1.26.4
blueprint